Clean up Router, add MainViewModel | #39
Background:

When the app first boots up, we need to determine if the user is or is not authenticated. This currently happens in the MainActivity, but it should really happen in a MainViewModel.

It was also noticed that using underscores in package names is not suggested, which is what we use for long screen packages.

There are some other minor style issues noted.

Changes:

Add `MainViewModel`, and move logic for determining authentication there.

Rename any package with an underscore to camel-case style.

Clean up small style issues.

Closes #39