Add crude login funct, connect login / register / welcome | #11 #8  #38
Background:

Currently, only registration is possible in the app, with no ability to login or out once authenticated / unauthenticated.

Changes:

Add a crude implementation to `LoginScreen` to allow the user to login to their existing account, or to click to register a new account.

Add a button to `RegisterScreen` to allow the user to log out of an existing account.

Testing:

Open app. Attempt to login to an existing account correctly / incorrect, register a new account, log out.

Closes #11, #8, #38