Crude forgot password implementation | #7 #9
Background:

It is important for a user to be able to reset their password from the application if necessary. There is currently a button on the login screen to reset, but it has no function.

Changes

This commit adds a functional, crude forgot password flow to the application. The user will click forgot password and enter an email. If the email is valid or not valid they are informed of that. They must then navigate to their e-mail and reset it from the link provided.

Closes #7 #9