refactor: Update UseCases & Repository to improve abstraction | #53
refactor: Update UseCases & Repository to improve abstraction | #53

Background:

We are adding in functionality to edit the username and password of an account. Ahead of this work, it was important to clean up our main interface which interacts with our authentication database, `AuthRepository`. This commit cleans up our generic response for receiving either a `Success` or `Error` from the database when that is all the information we want. (i.e., we don't want that to wrap around a piece of data).

Changes:

* Add `GetUserUseCase`, to abstract away the need for our individual viewmodels to directly invoke the `getCurrentUser` function within `AuthRepository`.
* Add `User`, a domain-layer interface specifying an object that contains only the information our ViewModels need, the username and email.
* Change forgot password response and any response where we don't want any data returned to be `Response`, but to wrap the `Unit`  class, which is also considered void.
* Rename `Resource` to `Response`

Testing:

Run the application, it should work exactly as it had previously with no changes.

Related to #53