feat: Add functional impl to edit username and password | #53
feat: Add functional impl to edit username and password | #53

Background:

Currently, though the navigation exists, it is not possible to edit your username or password - the confirm button had no functionality. This commit adds functionality to both edit username and edit password, allowing the user to do both in the application.

Changes:

* Add `UserUtils`, a utility file we can inject whenever we need to verify if an email or password is valid. The function within is placeholder for now.
* Change `EditPassScreen` to contain two text fields, prompting the user to input their old password in addition to a new password.
* Update `AuthRepository` and its impl to reauthenticate by accepting an old password parameter.
* When either property has been successfully set, inform the user.

Testing:

Open the app, click "Edit Profile" on the welcome screen, edit both the username and password - ensure they are updated.

Note: there is currently a visual "bug" where when you navigate up after changing the username it will not be reflected. This is not exactly a bug and is present on many applications when you make an edit due to the nature of how those updates would be made - but we can try to adjust this later.

Related to #53