feat: Lock go previous button and flip
* Added a property to `SwipeableCardState`, `isDragging` which reads true when card is dragged and reads false when card either animates back or snaps back.
* isDragging currently effects:
* Go previous IconButton and icon alpha.
* `flippable()` now takes `isDragging` as a enable value.
  * Check this for bugs as I did it super quick.

**Notable Issue**
Currently when a card is swiped away, there is a short period of time where you can click the card and a super weird flip effect happens.  I put an `isDraggingReset()` call inside the `sideEffect` but the issue seems to persist so see if you can take a look at it. If not no worries I'm gonna keep playing with it.